- Fix for a potential error in 5+ bone weight corrections due to errors with treating bytes as signed values. - Fix for creating modpacks that included mods past the 2GB limit.

+++ xivModdingFramework/Mods/FileTypes/TTMP.cs
@@ -238,7 +238,7 @@ namespace xivModdingFramework.Mods.FileTypes
                                 }
                             };
 
-                            var rawData = dat.GetRawData((int) simpleModData.ModOffset,
+                            var rawData = dat.GetRawData(simpleModData.ModOffset,
                                 XivDataFiles.GetXivDataFile(simpleModData.DatFile),
                                 simpleModData.ModSize);
